converted to float

+++ src/AnalogueInput.cpp
@@ -7,26 +7,26 @@
 AnalogueInput::AnalogueInput() {
     aDCHandler.initializeADC();
 }
-float AnalogueInput::turbineCurrentCapacity() { //Read and return turbine capacity current
+uint16_t AnalogueInput::turbineCurrentCapacity() { //Read and return turbine capacity current
     //Read Pin A2 using ADC
     aDCHandler.getVoltage(2);
     return ADCVoltage; //Amps 1:1 mapping
 }
 
-float AnalogueInput::pvCurrentCapacity() { //Read and return the PV Capacity current
+uint16_t AnalogueInput::pvCurrentCapacity() { //Read and return the PV Capacity current
     //Read Pin A3 using ADC
     aDCHandler.getVoltage(3);
     return ADCVoltage; //Amps 1:1 mapping
 }
 
-float AnalogueInput::busbarCurrent() {
+uint16_t AnalogueInput::busbarCurrent() {
     //Read Pin A1 using ADC
     aDCHandler.getVoltage(1);
     return ADCVoltage; //Amps 1:1 mapping
 }
 
-float AnalogueInput::busbarVoltage() {
+uint16_t AnalogueInput::busbarVoltage() {
     //Read Pin A0 using ADC (scaled down from 10v)
     aDCHandler.getVoltage(0);
-    return 100.0f * ADCVoltage; //Volts
+    return 100 * ADCVoltage; //Volts
 }
